+++ src/de/gwvprojekt/ui/GameUI.java
@@ -97,9 +97,9 @@ public class GameUI
 	public void updateUI()
 	{
 		int a = 0;
-		for (int x = 0; x < 4; x++)
+		for (byte x = 0; x < 4; x++)
 		{
-			for (int y = 0; y < 4; y++)
+			for (byte y = 0; y < 4; y++)
 			{
 				String text;
 				if (_data.getStringValue(x, y).equals("16"))
